producer/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
kafka-python==2.0.2
packaging==24.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
